requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.5
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
